buscador_normativo_mvp/requirements.txt
streamlit==1.37.1
pandas==2.2.2
python-dateutil>=2.8.2